+++ main.py
@@ -1,14 +1,9 @@
-users:list[dict]=[
-    {'name':'Martyna','location':'Zabki',"posts":5,},
-    {'name':'Wiktoria','location':'Radzymin',"posts":10,},
-    {'name':'Daria','location':'Losice',"posts":15,}
-]
+from utils.controller import get_user_info
+from utils.model import users
 
 
-def get_user_info(users_data:list[dict])->None:
-    for user in users_data:
-        print(f'Twoj znajomy: {user["name"]} z {user["location"]} opublikowal {user["posts"]}')
+def main():
+    get_user_info(users)
 
-get_user_info(users)
-
-
+if __name__ == '__main__':
+    main()
