+++ utils/model.py
@@ -1,0 +1,5 @@
+users:list[dict]=[
+    {'name':'Martyna','location':'Zabki',"posts":5,},
+    {'name':'Wiktoria','location':'Radzymin',"posts":10,},
+    {'name':'Daria','location':'Losice',"posts":15,}
+]
